+++ tests/en_test.go
@@ -55,8 +55,7 @@ func TestParseEn(t *testing.T) {
 
 	for _, tc := range cases {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
-			timeago.ClearCache()
-			timeago.Configure(timeago.Config{Language: langEn})
+			timeago.Reconfigure(timeago.Config{Language: langEn})
 
 			res, err := timeago.Parse(tc.date)
 
@@ -91,8 +90,7 @@ func TestParseEnWithSeconds(t *testing.T) {
 
 	for _, tc := range cases {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
-			timeago.ClearCache()
-			timeago.Configure(timeago.Config{Language: langEn})
+			timeago.Reconfigure(timeago.Config{Language: langEn})
 
 			res, err := timeago.Parse(tc.date)