+++ tests/no_suffix_test.go
@@ -65,8 +65,7 @@ func TestParseWithNoSuffixFlag(t *testing.T) {
 
 	for _, tc := range cases {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
-			timeago.ClearCache()
-			timeago.Configure(timeago.Config{Language: tc.lang})
+			timeago.Reconfigure(timeago.Config{Language: tc.lang})
 
 			res, err := timeago.Parse(tc.date, "noSuffix")