+++ tests/online_test.go
@@ -63,8 +63,7 @@ func TestParseWithOnlineFlag(t *testing.T) {
 
 	for _, tc := range cases {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
-			timeago.ClearCache()
-			timeago.Configure(timeago.Config{Language: langEn})
+			timeago.Reconfigure(timeago.Config{Language: langEn})
 
 			res, err := timeago.Parse(tc.date, "online")