+++ tests/ru_test.go
@@ -56,8 +56,7 @@ func TestParseRu(t *testing.T) {
 
 	for _, tc := range cases {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
-			timeago.ClearCache()
-			timeago.Configure(timeago.Config{Language: langRu})
+			timeago.Reconfigure(timeago.Config{Language: langRu})
 
 			res, err := timeago.Parse(tc.date)
 
@@ -93,8 +92,7 @@ func TestParseRuWithSeconds(t *testing.T) {
 
 	for _, tc := range cases {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
-			timeago.ClearCache()
-			timeago.Configure(timeago.Config{Language: langRu})
+			timeago.Reconfigure(timeago.Config{Language: langRu})
 
 			res, err := timeago.Parse(tc.date)