+++ timeago_test.go
@@ -24,8 +24,7 @@ func TestParseFunctionCanExceptTimestamp(t *testing.T) {
 		{getTimestampOfPastDate(time.Hour * 24 * 7), "1 week ago"},
 	}
 
-	ClearCache()
-	Configure(Config{Language: "en"})
+	Reconfigure(Config{Language: "en"})
 
 	for _, tc := range cases {
 		t.Run(tc.res, func(t *testing.T) {
@@ -60,8 +59,7 @@ func TestParseFunctionCanExceptTimePackage(t *testing.T) {
 		{time.Now().Add(-time.Hour * 11), "11 hours ago"},
 	}
 
-	ClearCache()
-	Configure(Config{Language: "en"})
+	Reconfigure(Config{Language: "en"})
 
 	for _, tc := range cases {
 		t.Run("Test for date "+tc.date.String(), func(t *testing.T) {
@@ -91,8 +89,7 @@ func TestParseFuncWillCalculateIntervalToFutureDate(t *testing.T) {
 		{time.Now().Add(time.Hour * 48), "2 days"},
 	}
 
-	ClearCache()
-	Configure(Config{Language: "en"})
+	Reconfigure(Config{Language: "en"})
 
 	for _, tc := range testCases {
 		t.Run("Test for date: "+tc.date.String(), func(t *testing.T) {